requirements.txt
# Game engine
catanatron>=0.6.0

# Our utilities (install from local path)
-e ../llm-game-utils

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Analysis
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0

# Optional: for better visualizations
plotly>=5.14.0
